8/20 - fix flow, enc init

+++ Marlin/src/module/planner.cpp
@@ -176,8 +176,8 @@ float Planner::mm_per_step[DISTINCT_AXES];      // (mm) Millimeters per step
 #endif
 
 #if HAS_EXTRUDERS
-  int16_t Planner::flow_percentage[EXTRUDERS] = ARRAY_BY_EXTRUDERS1(static_cast<int16_t>(100)); // Extrusion factor for each extruder
-  float Planner::e_factor[EXTRUDERS] = ARRAY_BY_EXTRUDERS1(1.0f); // The flow percentage and volumetric multiplier combine to scale E movement
+  uint16_t Planner::flow_percentage[EXTRUDERS] = ARRAY_BY_EXTRUDERS1(100); // Extrusion factor for each extruder
+  float Planner::e_factor[EXTRUDERS] = ARRAY_BY_EXTRUDERS1(1.00f); // The flow percentage and volumetric multiplier combine to scale E movement
 #endif
 
 #if DISABLED(NO_VOLUMETRICS)
